Change aspectRatio and voteAverage types from Double to Float in Image

+++ core/domain/src/main/kotlin/com/koniukhov/cinecircle/core/domain/model/Image.kt
@@ -5,7 +5,7 @@ data class Image(
     val width: Int,
     val height: Int,
     val countryCode: String,
-    val aspectRatio: Double,
-    val voteAverage: Double,
+    val aspectRatio: Float,
+    val voteAverage: Float,
     val voteCount: Int
 )
